requirements.txt
azure-messaging-webpubsubservice==1.0.0b1
django==3.2
djangorestframework==3.12.4
django-environ==0.4.5
djongo==1.3.6
websockets==10.0